# backend/docs/swagger.yaml
basePath: /api
definitions:
  github_com_thienel_go-backend-template_internal_interface_api_dto.APIResult:
    properties:
      command_error:
        type: string
      command_uuid:
        type: string
      no_push:
        type: boolean
      push_error:
        type: string
      request_type:
        type: string
      status:
        additionalProperties:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.PushStatus'
        type: object
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceInfoRequest:
    properties:
      queries:
        items:
          type: string
        type: array
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceInstallProfileRequest:
    properties:
      profile_id:
        type: integer
    required:
    - profile_id
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceLockRequest:
    properties:
      message:
        example: Device is locked by IT
        type: string
      phone_number:
        example: "+84123456789"
        type: string
      pin:
        example: "123456"
        type: string
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceRemoveProfileRequest:
    properties:
      profile_identifier:
        type: string
    required:
    - profile_identifier
    type: object
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceRestartRequest:
    properties:
      notify_user:
        type: boolean
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceWipeRequest:
    properties:
      disallow_proximity_setup:
        type: boolean
      obliteration_behavior:
        enum:
        - Default
        - DoNotObliterate
        - ObliterateWithWarning
        - Always
        type: string
      pin:
        example: "123456"
        type: string
      preserve_data_plan:
        type: boolean
    type: object
...
  github_com_thienel_go-backend-template_internal_interface_api_dto.PushStatus:
    properties:
      command_error:
        type: string
      push_error:
        type: string
      push_result:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_pkg_response.APIResponse-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  ? github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult
  : properties:
      data:
        $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.APIResult'
      error:
        $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.Error'
      is_success:
        type: boolean
      message:
        type: string
    type: object
...
  github_com_thienel_go-backend-template_pkg_response.Error:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.FieldError'
        type: array
      message:
        type: string
    type: object
  github_com_thienel_go-backend-template_pkg_response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for THD iOS MDM Portal.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: THD iOS MDM Portal API
  version: "1.0"
paths:
...
  /api/v1/devices:
    get:
      description: Fetch devices with pagination, sorting and filtering
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List devices
      tags:
      - Devices
  /api/v1/devices/{id}:
    get:
      description: Fetch single device details
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get device by ID
      tags:
      - Devices
  /api/v1/devices/{id}/install-profile:
    post:
      consumes:
      - application/json
      description: Queues MDM install profile command for a specific profile
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Profile to install
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceInstallProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Install profile on device
      tags:
      - Device Actions
  /api/v1/devices/{id}/lock:
    post:
      consumes:
      - application/json
      description: Queues MDM device lock command with optional PIN, message, and
        phone number
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Lock options
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceLockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Lock device
      tags:
      - Device Actions
  /api/v1/devices/{id}/remove-profile:
    post:
      consumes:
      - application/json
      description: Queues MDM remove profile command
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Profile to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceRemoveProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Remove profile from device
      tags:
      - Device Actions
  /api/v1/devices/{id}/request-info:
    post:
      consumes:
      - application/json
      description: Queues MDM device information command to query device attributes
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Information queries
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Request device information
      tags:
      - Device Actions
  /api/v1/devices/{id}/restart:
    post:
      consumes:
      - application/json
      description: Queues MDM restart device command
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Restart options
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceRestartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Restart device
      tags:
      - Device Actions
  /api/v1/devices/{id}/shutdown:
    post:
      description: Queues MDM shutdown device command
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Shutdown device
      tags:
      - Device Actions
  /api/v1/devices/{id}/wipe:
    post:
      consumes:
      - application/json
      description: Queues MDM erase device command with optional parameters
      parameters:
      - description: Device ID (UDID)
        in: path
        name: id
        required: true
        type: string
      - description: Wipe options
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_thienel_go-backend-template_internal_interface_api_dto.DeviceWipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-github_com_thienel_go-backend-template_internal_interface_api_dto_APIResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_thienel_go-backend-template_pkg_response.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Wipe device
      tags:
      - Device Actions
  /api/v1/devices/export:
    get:
      description: Export devices to CSV or JSON
      parameters:
      - description: Format (csv or json)
        in: query
        name: format
        type: string
      produces:
      - text/csv
      responses: {}
      security:
      - BearerAuth: []
      summary: Export devices
      tags:
      - Devices
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...